chore: move renderEmail
