+++ server/src/modules/actions/emails.actions.ts
@@ -1,8 +1,5 @@
-import { TemplateName } from "shared/mailer";
 import { IEvent } from "shared/models/events/event.model";
 
-import { getEmailInfos } from "@/common/services/mailer/mailer";
-import { generateHtml } from "@/common/utils/emailsUtils";
 import { getDbCollection } from "@/common/utils/mongodbUtils";
 
 export function addEmail(
@@ -127,26 +124,6 @@ export async function unsubscribeUser(id: string) {
       },
     },
     { returnDocument: "after" }
-  );
-}
-
-export async function renderEmail(token: string) {
-  const event = await getDbCollection("events").findOne<IEvent>({
-    "payload.emails.token": token,
-    name: "bal_emails",
-  });
-  if (!event) {
-    return;
-  }
-  const email = event.payload.emails.find((e) => e.token === token);
-  if (!email) {
-    return;
-  }
-  const { templateName, payload } = email;
-  return generateHtml(
-    payload.recipient.email,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    getEmailInfos(templateName as TemplateName, payload as any)
   );
 }
